Refactor struct tags in Movie type
This commit adds the - struct tag directive to the CreatedAt Movie field
since its not useful to end users. This commit also adds the "omitempty"
struct tag directive to the Movie fields Year, Runtime, and Genres to
ensure that empty values of these fields are hidden (i.e. Year of 0
doesn't make sense if we don't know the Year of a Movie).