Add psql DB to app
This commit adds db config struct to application struct. It builds on
this by creating an openDB function to create a psql DB connection pool
using the same db config in our application struct.

This commit all updates the logging in our application struct by adding
an errorLogger and refactor the existing logger to be an infoLogger.