Set up Movie Model and add placeholder methods
This commit creates the MovieModel struct type which wraps the database
connection pool. We also created a Models struct which is a single
convenient container to hold all our database models to make it easier
to action on each model type. We also created the custom error type,
ErrRecordNotFound that will be used whenever a movie record does not
exist in our database.

We also added the CRUD method placeholders (i.e. Insert, Get, Update,
and Delete) for the MovieModel struct type.