Create sortColumn & sortDirection methods
This commit creates two methods on Filters, sortColumn and
sortDirection.

sortColumn checks that the Sort field provided by the
client matches one of the entries in the SortSafeList. If it does, then
it extracts the column name by stripping the leading hyphen character
(if it exists).

sortDirection returns the sort direction ("ASC" or "DESC") depending on
the prefix character of the Sort field.