Update rateLimiter to be client specific
This commit updates the rateLimiter middleware to be client specific.
However, the current iteration is naive in that we store every client
IP address as key in a map with a rate.Limiter as the corresponding
value. Doing so will cause the map to grow indefinitely and taking
up more and more resources with each new IP address. We will refactor
this later on.