Refactor rateLimiter
This commit refactors the rateLimiter middleware so that it deletes
clients/IP-address from the clients map that haven't been seen in the
last 3 minutes. This prevents the clients map from getting infinitely
large and hogging resources. We do this by creating a goroutine that
removes these "old" entries in the clients map once every 1 minute.