Add rate limiting to app config struct
In this commit we add a limiter struct inside the app.config struct
along with corresponding command-line flags so that we can configure
our rate limiting at runtime. We include flags for requests/second,
burst, and enabled/disabled (so that we can turn off rate limiting,
entirely if we want).