Update server start to intercept SIGINT & SIGTERM
Update the server starting function to intercept SIGINT and SIGTERM
signals, log that we caught such signals and exit the app with status
code 0 (success). We do not catch SIGQUIT, because we want to have
the ability to quickly execute a "non-graceful" shutdown via a keyboard.
We may refactor this in the future.