Create requireActivatedUser middleware
requireActivatedUser middleware func is used on the movies routes and
returns a 401 Uauthorized status code if the user is anonymous (i.e.,
not authenticated). It also checks that if the user is not anonymous
but they are not activated then it returns a 403 Forbidden status code
to the user. All else, the users are considered authorized to view the
movies routes.