Refactor authentication/isActive middleware
Refactor the requireActivatedUser middleware to be split into 2:
requireAuthenticatedUser, which checks that the user is authenticated (i
.e., they are not anonymous), and requireActivatedUser, which checks
that the user is Active and calls requireAuthenticatedUser before being
executed itself.

This is better and makes more sense given we shouldn't be
checking if a user is activated unless we know exactly who they are
(i.e.,  they are authenticated).